split test taskmanager.
